Subprocess w/o embedding
To use on systems w/o an R library.
