+++ src/Makevars
@@ -1,6 +1,13 @@
 
-rem: r.o
-	$(MAIN_LINK) -o $@ r.o $(LIBR) $(LIBS)
+tools: rem asciicastclient$(SHLIB_EXT)
+
+rem: r.o common.o
+	$(MAIN_LINK) -o $@ r.o common.o $(LIBR) $(LIBS)
+
+asciicastclient$(SHLIB_EXT): client.o common.o
+	$(SHLIB_LINK) -o asciicastclient$(SHLIB_EXT) \
+		client.o common.o $(PKG_LIBS) \
+		$(SHLIB_LIBADD) $(LIBR)
 
 clean:
-	@rm -f rem r.o
+	@rm -f rem r.o client.o asciicastclient$(SHLIB_EXT)
